Module Expert.Serializer

type 'number serializer := Faraday.t -> 'number -> unit

val create : 'number serializer -> 'number t -> Faraday.t -> unit

Faraday serializer for a json object given a serializer for the number type. This minimizes whitespace in the serialized representation.

val create_hum : spaces:int -> 'number serializer -> 'number t -> Faraday.t -> unit

create_hum ~spaces is like serialize but it adds newlines and spaces in the same way that JSON.stringify(_,_,spaces) does in JavaScript. If spaces is 0 then the result is the same as that of serialize.