Module Jsonaf_kernel.Serializer

val serialize : t -> Faraday.t -> unit

Serialize a json object without any unnecessary whitespace.

val run : t -> string

Convert a json object to a string as serialize would.

val serialize_hum : spaces:int -> t -> Faraday.t -> unit

serialize_hum ~spaces t serializes t in a “human readable” form with newlines separating elements of objects and arrays and an additional indentation of spaces spaces for their contents. A typical choice is ~spaces:2.

Passing ~spaces:0 gives the exact same output as serialize.

val run_hum : spaces:int -> t -> string

run_hum ~spaces, t should produce very similar output to JSON.stringify(t,null,spaces) in JavaScript.